requirements.txt
jenkinsapi
python-jenkins